Remove custom setup end-point in Strapi
