+++ nuxt/server/api/flixes/setup.get.ts
@@ -7,7 +7,7 @@ export default defineEventHandler<Promise<Flix | void>>(async event => {
   const headers = generateHeaders(event);
 
   if (headers['x-flix']) {
-    return $fetch<Flix>(`${backendUrl}/setup`, {
+    return $fetch<Flix>(`${backendUrl}/flixes`, {
       headers,
     });
   }
