--- strapi/src/api/setup/controllers/setup.js
@@ -1,11 +1,0 @@
-'use strict';
-
-module.exports = {
-  retrieveSetup: async ctx => {
-    try {
-      ctx.body = ctx.state.flix;
-    } catch (err) {
-      ctx.body = err;
-    }
-  }
-};
