--- strapi/src/api/setup/routes/setup.js
@@ -1,13 +1,0 @@
-module.exports = {
-  routes: [
-    {
-      method: 'GET',
-      path: '/setup',
-      handler: 'setup.retrieveSetup',
-      config: {
-        policies: [],
-        middlewares: ['global::context'],
-      },
-    },
-  ],
-};
